Clicking link on "No Update" dialog closes the dialog
For unknown reason, after clicking "Go to download page" link, the dialog sinks on the bottom of Z-buffer of all application - user has to minimize or close all other programs (with GUI) to get back this dialog. This commit provides the (workaround) solution to close the dialog after clicking the link.